fix: clean up spinbike-ui clippy debt surfaced by the new CI gate
Newly-enabled `cargo clippy --manifest-path spinbike-ui/Cargo.toml`
(part of #122) surfaced 44 pre-existing warnings across 19 files —
accumulated debt from clippy never having run on this workspace
before. All fixes are mechanical, applying clippy's own suggestions
verbatim, no behavior change:

- `view! {}.into_any()` -> `().into_any()` (unused_unit / unit_arg) —
  login_link_form.rs, card_panel.rs, dashboard/mod.rs, login.rs,
  my_balance.rs
- `X.clone()` -> `X` where X: Copy (Callback<T> is Copy) —
  class_card.rs, segmented.rs, sheet.rs, edit_info_form.rs,
  users_by_movement.rs; two now-redundant `let x = x;` rebinds removed
  entirely (class_card.rs, segmented.rs) rather than left as no-op copies
- Nested `if`/`match` collapsed via let-chains (edition 2024) —
  api.rs, date_input.rs, install_prompt.rs (x2), activity_feed.rs,
  ws.rs (single_match -> if let)
- `*d = *d - X` / `*d = *d + X` -> `*d -= X` / `*d += X` —
  date_input.rs, reports/mod.rs
- Unnecessary `as u32` cast (get_full_year() already returns u32) —
  schedule.rs, staff_dashboard.rs
- Dead `let kind = s.kind.clone();` removed — admin.rs
- Redundant `let credit_val = credit_val;` removed (Copy type, no
  rebind needed to move into closure) — dashboard/mod.rs
- `wasm_bindgen::prelude::*` import gated `#[cfg(not(test))]` to match
  its only use site — lib.rs
- `.map(|e| render_row(e))` -> `.map(render_row)` (redundant_closure) —
  activity_feed.rs

[no-test: mechanical clippy-suggested rewrites with no behavior change
(unit-arg equivalences, Copy-type clone removal, if/match collapsing
via let-chains, compound-assign rewrites, dead-code removal); existing
E2E/unit tests cover the affected pages/components and continue to
pass, which is the regression guard here]

+++ spinbike-ui/src/pages/schedule.rs
@@ -26,7 +26,7 @@ pub struct ClassSlot {
 /// Compute current week (Mon-Sun) as Vec<(year, month, day)> and date strings.
 fn current_week_dates() -> (Vec<(i32, u32, u32)>, Vec<String>) {
     let now = js_sys::Date::new_0();
-    let year = now.get_full_year() as u32;
+    let year = now.get_full_year();
     let month = now.get_month() as i32; // 0-based
     let day = now.get_date() as i32;
     let dow = now.get_day(); // 0=Sun
@@ -39,7 +39,7 @@ fn current_week_dates() -> (Vec<(i32, u32, u32)>, Vec<String>) {
     let mut date_strs = Vec::with_capacity(7);
     for i in 0..7i32 {
         let d = js_sys::Date::new_with_year_month_day(
-            monday.get_full_year() as u32,
+            monday.get_full_year(),
             monday.get_month() as i32,
             monday.get_date() as i32 + i,
         );
